Improve profiling output.
Change the #AT() message to include the current time and the time elapsed
since the previous message, so that I don't need to use a calculator to
determine how much time has elapsed.

Insert more Profiler.At() calls so that we can see more of the steps involved,
specifically DataContext creation, QueryProvider calls (to e.g. get the
results of a query (QueryProvider<T>.Execute()) or create an expression tree
(QueryProvider<T>.CreateQuery<E>()), and see when DataContext.GetTable() is
invoked.