Fix DataContextTest.Ctor_ConnectionString_ExtraParameters_Munging() under Mono.
The problem is that Mono's bundled Npgsql implementation is so old that
NpgsqlConnection doesn't extend DbConnection, it instead implements
IDbConnection.  This causes DataContext.Connection to return null, causing the
test to fail with a NullReferenceException.  This is, of course, bad.

The fix is to use Mono.Data.Sqlite, which is more recent and continues to pass
under .NET.