Get ReadTest.C27_SelectEntitySet() working.
The problem is ExpressionDispatcher.RegisterAssociation() -- once called, it
causes a SQL JOIN to be used.  For example, given:

	from e in db.Employees orderby e.EmployeeID select new {e.Orders};

the following SQL would be created:

	SELECT e$.[EmployeeID]
	FROM [dbo].[Employees] AS e$
	 LEFT JOIN [dbo].[Orders] AS t1$ ON t1$.[EmployeeID] = e$.[EmployeeID]
	ORDER BY e$.[EmployeeID]

Note the 'LEFT JOIN'.

Furthermore, most tests require this behavior, but the LINQ query used used in
C27_SelectEntitySet() must NOT have the JOIN in order to function properly.

The fix (working so far) is to delay the JOIN creation unless it's needed, and
it's deemed as "needed" if EntitySetExpression.TableExpression is accessed.
Otherwise, no JOIN is needed, the JOIN is skipped, and things behave properly
for the C27_SelectEntitySet() LINQ case.

Yay.