Fix storage and member aliasing.
(Mutable global data is bad, mkay?)

Because theoriginal code was updating `association.Storage` and
`association.Member` (the "global data" referred to above), this was
*reducing* conflicts.  Unfortunately, this is bad because it caused other
sections of the code to not use the *original* values (as the original values
were lost), thus resulting in broken aliases.

Patch thanks to andershol.

Fixes issue 196.